@@ -1,38 +1,40 @@
-{
-  "name": "angular2",
-  "version": "1.0.0",
-  "scripts": {
-    "start": "tsc && concurrently \"tsc -w\" \"lite-server\" ",
-    "lite": "lite-server",
-    "postinstall": "typings install",
-    "tsc": "tsc",
-    "tsc:w": "tsc -w",
-    "typings": "typings"
-  },
-  "license": "ISC",
-  "dependencies": {
-    "@angular/common": "~2.0.1",
-    "@angular/compiler": "~2.0.1",
-    "@angular/core": "~2.0.1",
-    "@angular/forms": "~2.0.1",
-    "@angular/http": "~2.0.1",
-    "@angular/platform-browser": "~2.0.1",
-    "@angular/platform-browser-dynamic": "~2.0.1",
-    "@angular/router": "~3.0.1",
-    "@angular/upgrade": "~2.0.1",
-    "angular-in-memory-web-api": "~0.1.1",
-    "bootstrap": "^3.3.7",
-    "core-js": "^2.4.1",
-    "ng2-bootstrap": "^1.1.6",
-    "reflect-metadata": "^0.1.8",
-    "rxjs": "5.0.0-beta.12",
-    "systemjs": "0.19.39",
-    "zone.js": "^0.6.25"
-  },
-  "devDependencies": {
-    "concurrently": "^3.0.0",
-    "lite-server": "^2.2.2",
-    "typescript": "^2.0.3",
-    "typings": "^1.4.0"
-  }
-}
+    {
+      "name": "angular-quickstart",
+      "version": "1.0.0",
+      "scripts": {
+        "start": "concurrently \"npm run tsc:w\" \"npm run lite\" ",
+        "lite": "lite-server",
+        "tsc": "tsc",
+        "tsc:w": "tsc -w"
+      },
+      "licenses": [
+        {
+          "type": "MIT",
+          "url": "https://github.com/angular/angular.io/blob/master/LICENSE"
+        }
+      ],
+      "dependencies": {
+        "@angular/common": "~2.1.1",
+        "@angular/compiler": "~2.1.1",
+        "@angular/core": "~2.1.1",
+        "@angular/forms": "~2.1.1",
+        "@angular/http": "~2.1.1",
+        "@angular/platform-browser": "~2.1.1",
+        "@angular/platform-browser-dynamic": "~2.1.1",
+        "@angular/router": "~3.1.1",
+        "@angular/upgrade": "~2.1.1",
+        "angular-in-memory-web-api": "~0.1.13",
+        "core-js": "^2.4.1",
+        "reflect-metadata": "^0.1.8",
+        "rxjs": "5.0.0-beta.12",
+        "systemjs": "0.19.39",
+        "zone.js": "^0.6.25"
+      },
+      "devDependencies": {
+        "@types/core-js": "^0.9.34",
+        "@types/node": "^6.0.45",
+        "concurrently": "^3.0.0",
+        "lite-server": "^2.2.2",
+        "typescript": "^2.0.3"
+      }
+    }
