@@ -2,8 +2,10 @@
 import { NgModule }      from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule }   from '@angular/forms';
-import { HttpModule }    from '@angular/http';
+import { HttpModule}    from '@angular/http';
+
 import {} from '@angular/core/testing';
+
 
 //Pages
 import { AppComponent }             from './app.component';
@@ -13,6 +15,7 @@ import { LoginComponent }           from './login.component/login.component';
 import { SignUpComponent }          from './sign-up.component/sign-up.component';
 import { SettingsComponent }        from './settings.component/settings.component';
 import { FitnessComponent }        from './fitness.component/fitness.component';
+import { FitnessDetailComponent }       from './fitness.component/fitness-detail.component/fitness-detail.component';
 
 //Services
 import { routing }          from './app.routing/app.routing';
@@ -22,7 +25,6 @@ import { UserService } from './testservice/user.service';
 import { AuthenticationService } from './login.component/login.service';
 //form services
 import { FitnessService } from './fitness.component/fitness.service';
-
 
 @NgModule({
     imports: [
@@ -39,6 +41,7 @@ import { FitnessService } from './fitness.component/fitness.service';
         SignUpComponent,
         SettingsComponent,
         FitnessComponent,
+        FitnessDetailComponent,
     ],
     providers: [
         AuthGuard,
