@@ -11,8 +11,7 @@ import { LoginComponent }           from '../login.component/login.component';
 import { SignUpComponent }          from '../sign-up.component/sign-up.component';
 import { SettingsComponent }        from '../settings.component/settings.component';
 import { FitnessComponent }        from '../fitness.component/fitness.component';
-
-
+import { FitnessDetailComponent }       from '../fitness.component/fitness-detail.component/fitness-detail.component';
 
 const appRoutes: Routes = [
     { path: 'login', component: LoginComponent },
@@ -22,7 +21,7 @@ const appRoutes: Routes = [
     { path: 'settings', component: SettingsComponent, canActivate: [AuthGuard] },
     { path: 'signup', component: SignUpComponent, canActivate: [AuthGuard] },
     { path: 'fitnesstowork', component: FitnessComponent, canActivate: [AuthGuard] },
- 
+    { path: 'fitnessform/:_id',component: FitnessDetailComponent},
     // otherwise redirect to home
     { path: '**', redirectTo: '' }
     
