@@ -1,14 +1,18 @@
 export class Fitness {
-    q1: string;
-    q2: string;
-    q3: string;
+    _id: string;
+    email:string;
+    q1EmpName: string;
+    q2DateOfAssessment: string;
+    q3ReasonForAssessment: string;
     q4: string;
+    qMed: string;
     q5: string;
     q6: string;
     q7: string;
     q8: string;
     q9: string;
-    q10: string;
-    q11: string;
-    q12: string;
+    q10ActionTaken: string;
+    q11OwnerManager: string;
+    q12DateSigned: string;
+    __v:           number;
 }
